moves sandbox related code to enterprise
