+++ frontend/src/metabase/admin/permissions/selectors/confirmations.ts
@@ -8,7 +8,6 @@ import {
 import {
   getFieldsPermission,
   getSchemasPermission,
-  hasPermissionValueInGraph,
 } from "metabase/admin/permissions/utils/graph";
 import type Database from "metabase-lib/v1/metadata/Database";
 import type {
@@ -136,31 +135,6 @@ export function getControlledDatabaseWarningModal(
     return {
       title: t`Change access to this database to “Granular”?`,
       message: t`Just letting you know that changing the permission setting on this ${entityType} will also update the database permission setting to “Granular” to reflect that some of the database’s ${entityTypePlural} have different permission settings.`,
-      confirmButtonText: t`Change`,
-      cancelButtonText: t`Cancel`,
-    };
-  }
-}
-
-export function getSandboxedTableWarningModal(
-  permissions: GroupsPermissions,
-  groupId: number,
-  entityId: EntityId,
-  value: DataPermissionValue,
-) {
-  // if the user is sandboxing the table while there is create-queries permissions set to
-  // query builder and native for that group's access to the database being modified, we
-  // should prompt them that we will have to remove native access for all tables/schemas
-  if (
-    value === DataPermissionValue.SANDBOXED &&
-    hasPermissionValueInGraph(
-      permissions[groupId][entityId.databaseId],
-      DataPermissionValue.QUERY_BUILDER_AND_NATIVE,
-    )
-  ) {
-    return {
-      title: t`Change access to this database to “Sandboxed”?`,
-      message: t`As part of providing sandboxing we will also have to remove this group's native querying permissions from all tables and schemas in this database.`,
       confirmButtonText: t`Change`,
       cancelButtonText: t`Cancel`,
     };
